requirements.txt
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.0
branca==0.3.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
Django==1.11.20
django-crispy-forms==1.7.2
dlib==19.17.0
enum34==1.1.6
face-recognition==1.2.3
face-recognition-models==0.3.0
folium==0.8.3
gyp==0.1
html5lib==0.999999999
idna==2.8
Jinja2==2.10
lxml==4.2.1
MarkupSafe==1.1.1
matplotlib==2.1.1
mysqlclient==1.4.2.post1
numpy==1.16.2
olefile==0.45.1
PAM==0.4.2
pandas==0.24.2
pexpect==4.2.1
Pillow==5.1.0
pipenv==2018.11.26
psutil==5.4.2
pycairo==1.16.2
pygobject==3.26.1
pyinotify==0.9.6
PyMySQL==0.9.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
requests==2.9.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.11.0
subprocess32==3.2.7
typing==3.6.6
urllib3==1.24.1
virtualenv==16.4.3
virtualenv-clone==0.5.3
webencodings==0.5
xlrd==1.2.0
